software/apps/analytics/face_recognition/requirements.txt
numpy==1.15
tensorflow-gpu==1.12
scipy==1.2.1
scikit-learn==0.20.3
h5py
matplotlib==3.0.3
Pillow
requests
psutil
keras
imageio
